+++ php7/after.sh
@@ -4,13 +4,24 @@
 # add any commands you wish to this file and they will
 # be run after the Homestead machine is provisioned.
 
+if [ -e /.installed ]; then
+    echo 'Already installed.'
 
-# update
-sudo apt-get update
+else
+    echo ''
+    echo 'INSTALLING'
+    echo '----------'
 
-# install Ruby and mailcatcher
-sudo apt-get -y install ruby ruby-dev
-sudo gem install mailcatcher
+    # update
+    sudo apt-get update
+
+    # install Ruby and mailcatcher
+    sudo apt-get -y install ruby ruby-dev
+    sudo gem install mailcatcher
+
+    touch /.installed
+
+fi
 
 
 ## show versions
@@ -26,5 +37,7 @@ echo 'Homestead Box' \
 && echo 'npm' `npm -v` \
 && mailcatcher --version > homestead.box.txt
 
+echo ''
+echo 'BOX INFO'
 cat homestead.box.txt
-
+echo '----------'
